% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp2}
\alias{exp2}
\title{Stability related substance data of three batches}
\format{
A data frame with 48 observations and 3 variables:
\describe{
  \item{Batch}{Factor with levels \code{b4}, \code{b5} and \code{b8}}
  \item{Month}{Numeric representing the time points of testing from the
               start (0 months) to the end (24 months) of the study.}
  \item{Related}{Numeric of the measured related substance level in \%LC}
}
}
\source{
\href{http://www.ivtnetwork.com/sites/default/files/LinearRegression.pdf}{
      LeBlond \emph{et al}., \emph{J Valid Technol} 2011}
}
\usage{
data(exp2)
}
\description{
A data set containing the related substance stability data (in \% of label
claim (LC)) of three batches obtained over a 24 months period of a drug
product. A total of \eqn{n = 24} independent measurements are available
(corresponding to data shown in Table XI in LeBlond \emph{et al}).
}
\examples{
\dontrun{exp2}
}
\references{
LeBlond, D., Griffith, D. and Aubuchon, K. Linear Regression 102: Stability
Shelf Life Estimation Using Analysis of Covariance. \emph{J Valid Technol}
(2011) \strong{17}(3): 47-68.
}
\keyword{datasets}
