% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp4}
\alias{exp4}
\title{Stability data of 4 batches}
\format{
A data frame with 36 observations and 3 variables:
\describe{
  \item{Batch}{Factor with levels \code{1_11}, \code{2_12}, \code{3_13} and
    \code{4_14}.}
  \item{Month}{Numeric representing the time points of testing from the
               start (0 months) to the end (24 months) of the study.}
  \item{Conc}{Numeric of the meausred concentrations in \eqn{mg / kg}.}
}
}
\source{
See reference: Example data set (Stability.jmp) used in chapter
\emph{Stability Analysis}, p. 174-176.
}
\usage{
data(exp4)
}
\description{
A data set containing the concentration data (mg/kg) of four batches
obtained over a 24 months period of a drug product. A total of \eqn{n = 36}
independent measurements are available (corresponding to the data set
\emph{Reliability/Stability.jmp} in \emph{JMP(R) 12 Reliability and Survival
Methods} manual).
}
\examples{
str(exp4)
}
\references{
SAS Institute Inc. 2015. \emph{JMP(R) 12 Reliability and Survival Methods}.
Cary, NC: SAS Institute Inc.
}
\keyword{datasets}
