% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{find_poi}
\alias{find_poi}
\title{Point of intersection}
\usage{
find_poi(
  srch_range,
  model,
  sl,
  alpha = 0.05,
  ivl = "confidence",
  ivl_type = "one.sided",
  ivl_side = "lower",
  ...
)
}
\arguments{
\item{srch_range}{A vector of length \code{2} specifying the end-points of
the (time) range within which the minimal distance is expected to be found.}

\item{model}{A linear model object of type \code{lm}.}

\item{sl}{A numeric variable specifying the \dQuote{specification limit}
(SL). Another kind of limit may be regarded as SL.}

\item{alpha}{A numeric value specifying the significance level of the
confidence or prediction interval that is calculated for the provided
linear model. The default value is \code{0.05}.}

\item{ivl}{A character string of either \code{"confidence"} or
\code{"prediction"}, i.e. specifying the type of interval of interest.
The default is \code{"confidence"}.}

\item{ivl_type}{A character string specifying if a \dQuote{one sided} or
a \dQuote{two sided} interval should be calculated, i.e. either
\code{"one.sided"} or \code{"two.sided"}, respectively. The default is
\code{"one.sided"}.}

\item{ivl_side}{A character string specifying if the \dQuote{upper} or the
\dQuote{lower} limit is the relevant limit, i.e. either \code{"upper"} or
\code{"lower"}, respectively. The default is \code{"lower"}.}

\item{...}{Additional named or unnamed arguments passed on to
\code{uniroot()}.}
}
\value{
A numeric value representing the value of \eqn{x} where the distance
between the two lines of interest is minimal is returned.
}
\description{
The function \code{find_poi()} determines the point where the distance
between two lines is minimal, e.g., the distance between a specification or
expiry limit and a confidence or prediction interval. The estimation is done
by aid of \code{uniroot()} from the \code{stats} package.
}
\details{
The function \code{find_poi()} (find the \sQuote{point of
intersection}) estimates the value of \eqn{x} (e.g., the time) where the
difference between the upper or lower confidence or prediction interval and
the upper or lower acceptance criterion (e.g., the specification or the
expiry limit) is minimal. Confidence or prediction intervals are calculated
for the \code{model} provided. The POI is determined by aid of the
\code{uniroot()} function from the \code{stats} package. The distance
between the two lines of interest is calculated using the function
\code{get_distance()}, and it is this distance which \code{uniroot()}
tries to minimise.

#@inheritSection get_distance How to estimate shelf life or expiry
}
\references{
Therapeutic Goods Administration (TGA) of the Department of Health of the
Australian Government, Australian Regulatory Guidelines for Prescription
Medicines (ARGPM), Stability testing for prescription medicines,
Version 1.1, March 2017  \cr
\href{https://www.tga.gov.au/stability-testing-prescription-medicines}{%
      ARGPM: Stability testing for prescription medicines}

International Conference on Harmonisation of Technical Requirements for
Registration of Pharmaceuticals for Human (ICH), Harmonised Tripartite
Guideline, Evaluation of Stability Data Q1E, step 4 version 2003
(CPMP/ICH/420/02).  \cr
\href{https://www.ich.org/page/quality-guidelines}{ICH Quality Guidelines}
}
\seealso{
\code{\link{get_distance}}, \code{\link[stats]{uniroot}},
\code{\link{expirest_osle}}, \code{\link{expirest_wisle}}.
}
\keyword{internal}
