% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_hlines}
\alias{get_hlines}
\title{Prepare horizontal lines}
\usage{
get_hlines(model)
}
\arguments{
\item{model}{An \sQuote{\code{expirest_osle}} or an
\sQuote{\code{expirest_Wisle}} object, i.e. a list returned
by the \code{\link{expirest_osle}()} or by the
\code{\link{expirest_wisle}()} function.}
}
\value{
A data frame with the columns \sQuote{Response}, Item, Colour and
Type is returned, where the column name \sQuote{Response} is a placeholder
for the corresponding variable name. The data frame has up to two rows
representing the specification limit(s).
}
\description{
The function \code{get_hlines()} prepares a data frame for putting
horizontal lines on a plot prepared by the \code{ggplot()} function from
the \sQuote{\code{ggplot2}} package.
}
\details{
The function \code{get_hlines()} expects various pieces
of information characterising an \sQuote{\code{expirest_osle}} or an
\sQuote{\code{expirest_wisle}} model. Based on the information provided,
the function prepares a data frame that that is used by the functions
\code{\link{plot_expirest_osle}()} or \code{\link{plot_expirest_wisle}()}
to put lines on the graph that is prepared by these functions.
}
\seealso{
\code{\link{plot_expirest_osle}}, \code{\link{plot_expirest_wisle}},
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_text}}.
}
\keyword{internal}
