% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_linear_models}
\alias{get_linear_models}
\title{Linear model fitting}
\usage{
get_linear_models(data, response_vbl, time_vbl, batch_vbl)
}
\arguments{
\item{data}{A data frame with the columns specified by \code{response_vbl},
\code{time_vbl} and \code{batch_vbl}.}

\item{response_vbl}{A character string specifying the response variable name
that must be a column of \code{data}.}

\item{time_vbl}{A character string specifying the time variable name that
must be a column of \code{data}.}

\item{batch_vbl}{A character string specifying the column in \code{data}
with the grouping information (i.e. a factorial variable) for the
differentiation of the observations of the different batches.}
}
\value{
A list of three elements is returned,
containing the following elements:
\item{Models}{A list of four elements named \code{cics}, \code{dics},
  \code{dids} and \code{individual}. The first three elements contain the
  \sQuote{\code{lm}} objects of the \dQuote{common intercept / common slope}
  (\code{cics}), \dQuote{different intercept / common slope} (\code{dics})
  and \dQuote{different intercept / different slope} (\code{dids}) models.
  The fourth element is a list of the \sQuote{\code{lm}} objects of the
  models obtained from fitting the data of each batch individually. The
  \code{cics}, \code{dics} and \code{dids} elements are \code{NA} if data
  of only a single batch were provided.}
\item{AIC}{A numeric named vector of the Akaike Information Criterion (AIC)
  values of the \code{cics}, \code{dics} and \code{dids} models.}
\item{BIC}{A numeric named vector of the Bayesian Information Criterion (BIC)
  values of each of the \code{cics}, \code{dics} and \code{dids} models.}
}
\description{
The function \code{get_linear_models()} fits four types of linear regression
models that are often used for the assessment of stability data, e.g. for
the estimation of the shelf life or retest period.
}
\details{
The function \code{get_linear_models()} expects a data frame with
a response variable, a time variable and a categorical variable which
usually has factor levels of multiple batches of a drug product that was
assessed over a certain period of time with respect to the time-dependen
behavior of characteristic parameters. Using these results, the function
fits
\itemize{
 \item a \emph{common intercept / common slope} model (cics),
 \item a \emph{different intercept / common slope} model (dics) or
 \item a \emph{different intercept / different slope} model (dids) and
 \item individual regression models
}

If the categorical variable has only a single factor level, then the first
three models are \code{NA} and only an individual regression model is fitted.
}
\seealso{
\code{\link{expirest_osle}}.
}
\keyword{internal}
