% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_n_list_levels}
\alias{get_n_list_levels}
\title{Determine the level of nesting of a list}
\usage{
get_n_list_levels(x)
}
\arguments{
\item{x}{A list.}
}
\value{
An integer representing the number of levels of the list. If an
object is passed on to \code{x} that is not a list \code{0} is returned.
}
\description{
The function \code{get_n_list_levels()} determines the number of levels of
a nested list.
}
\details{
The function \code{get_n_list_levels()} determines the number of
levels of a (nested) list.
}
\seealso{
\code{\link{extract_from_ll_wcsl}}, \code{\link{extract_wc_x}},
\code{\link{get_wisle_poi_list}}.
}
\keyword{internal}
