% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plot.plot_expirest_osle}
\alias{plot.plot_expirest_osle}
\title{Plot illustrating the shelf life estimation (osle)}
\usage{
\method{plot}{plot_expirest_osle}(x, ...)
}
\arguments{
\item{x}{An object of class \sQuote{\code{plot_expirest_osle}} returned by
the \code{\link{plot_expirest_osle}()} function.}

\item{...}{Further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{formatC}()} function.}
}
\value{
The \sQuote{\code{plot_expirest_osle}} object passed to the \code{x}
parameter is returned invisibly.
}
\description{
This is a method for the function \code{plot()} for objects of class
\sQuote{\code{plot_expirest_osle}}.
}
\details{
The element \code{Graph} of the \sQuote{\code{plot_expirest_osle}}
object that is returned by the function \code{\link{plot_expirest_osle}()}
is an object of class \sQuote{\code{ggplot}}, generated by the function
\code{\link[ggplot2]{ggplot}()} from the \sQuote{\code{ggplot2}} package.
Thus, the corresponding \code{plot} method is used for plotting. Arguments
to the \code{\link[ggplot2]{ggplot}()} function can be passed via the
\code{...} parameter.
}
\seealso{
\code{\link{expirest_osle}}, \code{\link{plot_expirest_osle}},
\code{\link[ggplot2]{ggplot}()}, \code{\link[utils]{methods}}.
}
