% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.expirest_osle}
\alias{print.expirest_osle}
\title{Print a summary of the shelf life estimation (osle)}
\usage{
\method{print}{expirest_osle}(x, ...)
}
\arguments{
\item{x}{an \code{expirest_osle} object, i.e. a list returned by the
\sQuote{expirest_osle()} function.}

\item{...}{further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{formatC}} function.}
}
\value{
The \code{expirest_osle} object passed to the \code{model} parameter
is returned invisibly.
}
\description{
This is a method for the function \code{print()} for objects of class
\code{expirest_osle}.
}
\details{
The function \code{print.expirest_osle()} prints the most relevant
information in an \code{expirest_osle} object.
}
\seealso{
\code{\link{expirest_osle}}, \code{\link{expirest_wisle}},
\code{\link[base]{formatC}}, \code{\link[utils]{methods}}.
}
