% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.expirest_wisle}
\alias{print.expirest_wisle}
\title{Print a summary of the what-if shelf life estimation (wisle)}
\usage{
\method{print}{expirest_wisle}(x, ...)
}
\arguments{
\item{x}{an \sQuote{\code{expirest_wisle}} object, i.e. a list returned by
the \code{expirest_wisle()} function.}

\item{...}{Further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{formatC}()} function.}
}
\value{
The \sQuote{\code{expirest_wisle}} object passed to the
\code{object} parameter is returned invisibly.
}
\description{
This is a method for the function \code{print()} for objects of class
\sQuote{\code{expirest_wisle}}.
}
\details{
The function \code{print.expirest_wisle()} prints the most relevant
information in an \sQuote{\code{expirest_wisle}} object.
}
\seealso{
\code{\link{expirest_wisle}}, \code{\link{expirest_osle}},
\code{\link[base]{formatC}}, \code{\link[utils]{methods}}.
}
