% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{try_get_model}
\alias{try_get_model}
\title{Try getting results from expression}
\usage{
try_get_model(expr)
}
\arguments{
\item{expr}{An expression to be executed.}
}
\value{
A list of length three with the following elements is returned
\item{Model}{The object returned by the expression, if applicable,
  or otherwise NULL}
\item{Warning}{A warning message, if applicable, or otherwise NULL}
\item{Error}{An error message, if applicable, or otherwise NULL}
}
\description{
The function \code{get_trial_list()} tries to execute the expression handed
over and returns a list of the outcome.
}
\details{
The function tries to execute the expression handed over. It case
of an error it does not stop execution but in any case returns a list as
specified in chapter \emph{Value}. If execution of the expression was
successful, the results are stored in element \emph{Model}. In case of
warnings or errors, the messages are silently redirected to the elements
\emph{Warning} or \emph{Error}, respectively.
}
\keyword{internal}
