% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expirest_wisle.R
\name{expirest_wisle}
\alias{expirest_wisle}
\title{What-if (approach for) shelf life estimation (wisle)}
\usage{
expirest_wisle(
  data,
  response_vbl,
  time_vbl,
  batch_vbl,
  rl,
  rl_sf,
  sl,
  sl_sf,
  srch_range,
  alpha = 0.05,
  alpha_pool = 0.25,
  xform = c("no", "no"),
  shift = c(0, 0),
  sf_option = "loose",
  ivl = "confidence",
  ivl_type = "one.sided",
  ivl_side = "lower",
  ...
)
}
\arguments{
\item{data}{A data frame with the columns specified by \code{response_vbl},
\code{time_vbl} and \code{batch_vbl}.}

\item{response_vbl}{A character string specifying the response variable name
that must be a column of \code{data}.}

\item{time_vbl}{A character string specifying the time variable name that
must be a column of \code{data}.}

\item{batch_vbl}{A character string specifying the column in \code{data}
with the grouping information (i.e. a factorial variable) for the
differentiation of the observations of the different batches.}

\item{rl}{A numeric value or a numeric vector specifying the release
specification limit(s) for which the corresponding expiry should be
estimated.}

\item{rl_sf}{A positive integer or a vector of positive integers specifying
the number of \dQuote{significant figures} (sf) of \code{rl}. It must have
the same length as \code{rl}.}

\item{sl}{A numeric value or a numeric vector of length \code{2} specifying
the specification limit or limits. If a vector is provided it must be of
the form \code{c(lower limit, upper limit)}.}

\item{sl_sf}{A positive integer or a vector of positive integers specifying
the number of \dQuote{significant figures} (sf) of \code{sl}. It must have
the same length as \code{sf}.}

\item{srch_range}{A vector of length \code{2} specifying the end-points of
the (time) range that is supposed to contain the shelf life or retest
period.}

\item{alpha}{A numeric value specifying the significance level for the
calculation of confidence or prediction intervals. The default value is
\code{0.05}.}

\item{alpha_pool}{A numeric value specifying the type I error rate for the
test of the poolability of the batches. The default value is \code{0.25},
following ICH Q1E guideline. The value should not be changed unless
supported by well-founded reasons.}

\item{xform}{A vector of two character strings specifying the transformation
of the response and the time variable. The default is \dQuote{no}
transformation, i.e. \code{c("no", "no")}, where the first element
specifies the transformation of the \eqn{x} variable and the second
element the transformation of the \eqn{y} variable. Valid alternatives
for \eqn{x} and/or \eqn{y} variable transformation are \code{"log"}
(natural logarithm), \code{"sqrt"} (square root) and \code{"sq"} (square).}

\item{shift}{A vector of two values which will be added to the variables
\eqn{x} and/or \eqn{y} before they are transformed as specified by the
\code{xform} parameter, where the first element will be added to the
\eqn{x} variable and the second element to the \eqn{y} variable. The
purpose is to prevent an undefined state which could arise when variables
with values of \eqn{\leq 0} are log or square root transformed. The
default is \code{c(0, 0)}. For transformation of the \eqn{x} variable
\eqn{1} should be chosen as shift parameter in case of \code{log} and
\eqn{0} in case of \code{sqrt} and \code{sq} transformation because
then the intercept will be at \eqn{x = 0}. For transformation of the
\eqn{y} variable any value can be chosen as shift parameter.}

\item{sf_option}{A character string specifying if the limits (\code{rl}
or \code{sl}) should be regarded as \dQuote{tight} or \dQuote{loose}, i.e.
either \code{"tight"} or \code{"loose"}, respectively. The option
\code{"tight"} means that the limits are rounded to the specified number
of significant figures specified by the parameters \code{rl_sf} and
\code{sl_sf}. In case of the option \code{"loose"} the limits are rounded
to the specified number of significant figures (\eqn{n}), followed by the
subtraction of \eqn{1} from the \eqn{n^{th}} digit and addition of
\eqn{5} to the \eqn{(n + 1)^{th}} digit if \code{ivl_side} is
\code{"lower"}, or followed by the addition of \eqn{4} to the
\eqn{(n + 1)^{th}} digit if \code{ivl_side} is \code{"upper"}.}

\item{ivl}{A character string of either \code{"confidence"} or
\code{"prediction"}, i.e. specifying the type of interval of interest.
The default is \code{"confidence"}.}

\item{ivl_type}{A character string specifying if a \dQuote{one sided} or
a \dQuote{two sided} interval should be calculated, i.e. either
\code{"one.sided"} or \code{"two.sided"}, respectively. The default is
\code{"one.sided"}.}

\item{ivl_side}{A character string specifying if the \dQuote{upper} or the
\dQuote{lower} limit is the relevant limit, i.e. either \code{"upper"} or
\code{"lower"}, respectively. The default is \code{"lower"}.}

\item{...}{Additional named or unnamed arguments passed on to
\code{uniroot()}.}
}
\value{
An object of class \sQuote{\code{expirest_wisle}} is returned,
containing the following list elements:
\item{Data}{Data frame of the original data including new columns with
  transformed variables, if applicable.}
\item{Parameters}{A list of the parameters with the elements \code{alpha},
  \code{alpha.pool}, \code{ivl}, \code{ivl.type} and \code{ivl.side}.}
\item{Variables}{A list of the variable names, i.e. the original names of
  the \code{batch_vbl}, the \code{time_vbl} and the \code{response_vbl} and,
  if applicable, of the transformed variables.}
\item{Model.Type}{A list of five elements specifying which model, based on
  the ANCOVA analysis, suits best. The first element (\code{type.spec})
  is a numeric vector of length 2 specifying the best model accepted at the
  significance level of 0.25. The first number represents the decision on
  the intercept and the second on the slope, where \code{1} stands for
  \dQuote{common} and \code{2} stands for \dQuote{different}. The second
  element (\code{type.acronym}) is an acronym representing the first item.
  The third to fifth elements contain the names of the model variables, i.e.
  \code{response.vbl}, \code{time.vbl} and \code{batch.vbl}.}
\item{Models}{A list of all possible models (i.e. \sQuote{\code{lm}}
  objects) that are relevant, i.e. \code{cics}, \code{dics} and \code{dids}.
  The second element of the list is a string summarising the information in
  the first element, i.e. either \code{cics}, \code{dics} or \code{dids}.}
\item{AIC}{A numeric named vector of the Akaike Information Criterion (AIC)
  values of each of the three fitted models.}
\item{BIC}{A numeric named vector of the Bayesian Information Criterion (BIC)
  values of each of the three fitted models.}
\item{wc.icpt}{A data frame of the worst case intercepts of each of the
  three fitted models.}
\item{wc.batch}{A list of numeric value(s) of the worst case batch(es) per
  model type.}
\item{Limits}{A list of all limits.}
\item{POI}{A data frame of the intercepts, the differences between release
  and shelf life limits, the WCSLs, the expiry and release specification
  limits, the shelf lives and POI values.}

The \code{POI} data frame has the following columns:
\item{Intercept.cics}{The intercept of the worst case batch for the cics
  model.}
\item{Intercept.dics}{The intercept of the worst case batch for the dics
  model.}
\item{Intercept.dids}{The intercept of the worst case batch for the dids
  model.}
\item{Delta.cics}{Absolute difference between the release and and the shelf
  life specification for the cics model.}
\item{Delta.dics}{Absolute difference between the release and and the shelf
  life specification for the dics model.}
\item{Delta.dids}{Absolute difference between the release and and the shelf
  life specification for the dids model.}
\item{WCSL.cics}{WCSL for the cics model.}
\item{WCSL.dics}{WCSL for the dics model.}
\item{WCSL.dids}{WCSL for the dids model.}
\item{Exp.Spec}{The (expiry) specification, i.e. the specification which is
  relevant for the determination of the expiry.}
\item{Rel.Spec}{The calculated release specification.}
\item{Shelf.Life.cics}{The estiamted shelf life for the cics model.}
\item{Shelf.Life.dics}{The estiamted shelf life for the dics model.}
\item{Shelf.Life.dids}{The estiamted shelf life for the dids model.}
\item{POI.Model.cics}{The POI  of the cics model.}
\item{POI.Model.dics}{The POI of the dics model.}
\item{POI.Model.dids}{The POI of the dids model.}
}
\description{
Based on a linear regression model fitted to a stability data set the
function \code{expirest_wisle()} estimates the expiry for the specified
release and specification limit following the ARGPM guidance
\dQuote{Stability testing for prescription medicines}. The abbreviation
\dQuote{wisle} stands for \dQuote{what-if shelf life estimation} (because
it estimates the shelf life (\dQuote{what}) for a given release limit
(\dQuote{if})).
}
\details{
For the shelf life estimation for submissions in Australia the
Australian Regulatory Guidelines for Prescription Medicines (ARGPM), i.e.
the guidance \dQuote{Stability testing for prescription medicines}, is
binding. In chapter 14.3.1, \dQuote{Predicting shelf life from stability
data}, it is described how the estimation should be done. It reads as
follows: \cr
Take into account the worst case situation at batch release. For example:
\describe{
 \item{1a)}{For medicine that has a lower Assay release limit of 95 per cent
   and a lower assay expiry limit of 90 per cent, the maximum decrease in
   assay allowed over the shelf-life is 5 per cent.}
 \item{2a)}{Similarly, for a medicine that has a release limit for an
   individual impurity of 0.2 per cent and an expiry limit of 0.5 per cent,
   the maximum increase in the impurity allowed over the shelf life is
   0.3 per cent.}
 \item{1b)}{For the same medicine, if the actual release assay result is
   101 per cent, then the shelf life should be determined at the time the
   medicine (or confidence interval of the regression line) decreases by
   5 per cent and reaches 96 per cent, rather than the expiry limit
   (90 per cent). This takes into account the possibility of batches being
   released at the lower release limit (i.e. 95 per cent) and ensures they
   will comply with the expiry limit throughout the shelf life.}
 \item{2b)}{Similarly, if the actual impurity content is 0.1 per cent at
   batch release, the shelf life should be determined as the time the
   95 per cent confidence interval reaches 0.4 per cent (i.e. increases by
   0.3 per cent).}
}

Therefore, according to the ARGPM guidance \dQuote{Stability testing for
prescription medicines}, it is necessary to define
\itemize{
 \item a release limit and
 \item an expiry limit or shelf life limit (usually the specification limit).
}
If both of these limits were the same the shelf life would be zero, i.e. no
change allowed. \cr

For the estimation of the shelf life or expiry limit it is necessary to find
the point where the upper or lower 95\% confidence interval limit of the
linear model fitted to the data (by aid of \code{lm}) intersects the
\dQuote{worst case scenario limit} (WCSL), which is defined by the ARGPM
guidance \dQuote{Stability testing for prescription medicines} as the
intercept \eqn{\pm} difference between the specification limit and the
release limit, where this differences is added (\eqn{+}) if the upper limit
is relevant or it is subtracted (\eqn{-}) if the lower limit is relevant.
In this package, this point is called the \dQuote{point of intersection} or
\dQuote{point of interest}, abbreviated POI.
}
\section{Checking batch poolability}{

According to ICH Q1E guideline, construction of the 95\% confidence interval
on the basis of the combined data of all test batches is allowed only if it
has been confirmed by aid of a statistical test whether the regression lines
from the different batches have a common slope and a common intercept. A
significance level of \code{alpha_pool = 0.25} has to be used for both
batch-related terms, and the test of the slopes has to precede the test of
the intercepts. From these tests, three possible models may be appropriate,
i.e.
\itemize{
 \item a \emph{common intercept / common slope} model (cics),
 \item a \emph{different intercept / common slope} model (dics) or
 \item a \emph{different intercept / different slope} model (dids).
}
The \emph{common intercept / different slope} model is not of practical
relevance because the corresponding model is missing an effect. If the slopes
are significantly different, there is no point comparing intercepts.

These requirements can be checked by aid of an \dQuote{ANalysis of
COVAriance} (ANCOVA) including the batch variable as interaction term. The
full ANCOVA model simultaneously tests all the effects, and non-significant
effects can be identified and removed for fitting of the final regression
model that is used for the estimation of the shelf life or retest period.

The significance level (\code{alpha_pool = 0.25}, Type I error) is used to
increase the power of the test to detect cases where the data should not be
pooled. Setting \code{alpha_pool = 0.25} decreases the probability of
incorrectly concluding that stability data from multiple batches can be
pooled. On the other hand, though, it increases the probability of using a
single batch to determine expiry when pooling batches would be more
appropriate.
}

\examples{
# Successful estimation
tmp <-
  expirest_wisle(data = exp1[exp1$Batch \%in\% c("b2", "b5", "b7"), ],
                 response_vbl = "Potency", time_vbl = "Month",
                 batch_vbl = "Batch", rl = 98, rl_sf = 3, sl = 95,
                 sl_sf = 3, srch_range = c(0, 500), alpha = 0.05,
                 alpha_pool = 0.25, xform = c("no", "no"), shift = c(0, 0),
                 sf_option = "tight", ivl = "confidence",
                 ivl_type = "one.sided", side = "lower")

# (Expected) results in tmp[["POI"]]
# Intercept.cics Intercept.dics Intercept.dids Delta.cics Delta.dics
#       100.5669       100.3638       100.2491          3          3
# Delta.dids WCSL.cics WCSL.dics WCSL.dids Exp.Spec.Report Exp.Spec
#          3  97.56688  97.36375  97.24914              95       95
# Rel.Spec.Report Rel.Spec Shelf.Life.cics Shelf.Life.dics
#              98       98        14.07398        13.23176
# Shelf.Life.dids POI.Model.cics POI.Model.dics POI.Model.dids
#        13.34561       25.99576       24.56688       23.14804

# Unsuccessful estimation
\dontrun{
  tmp <-
    expirest_wisle(data = exp1[exp1$Batch \%in\% c("b2", "b5", "b7"), ],
                   response_vbl = "Potency", time_vbl = "Month",
                   batch_vbl = "Batch", rl = 98, rl_sf = 3, sl = 95,
                   sl_sf = 3, srch_range = c(0, 500), alpha = 1E-16,
                   alpha_pool = 0.25, xform = c("no", "no"), shift = c(0, 0),
                   sf_option = "tight", ivl = "confidence",
                   ivl_type = "one.sided", side = "lower")
}

# (Expected) results in tmp[["POI"]]
# Intercept.cics Intercept.dics Intercept.dids Delta.cics Delta.dics
#             NA             NA             NA         NA         NA
# Delta.dids WCSL.cics WCSL.dics WCSL.dids Exp.Spec.Report Exp.Spec
#         NA        NA        NA        NA              95       95
# Rel.Spec.Report Rel.Spec Shelf.Life.cics Shelf.Life.dics
#              98       98              NA              NA
# Shelf.Life.dids POI.Model.cics POI.Model.dics POI.Model.dids
#              NA        14.2773       1.651248             NA
}
\references{
Therapeutic Goods Administration (TGA) of the Department of Health of the
Australian Government, Australian Regulatory Guidelines for Prescription
Medicines (ARGPM), Stability testing for prescription medicines,
Version 1.1, March 2017\cr
\url{https://www.tga.gov.au/stability-testing-prescription-medicines}

International Council for Harmonisation of Technical Requirements for
Registration of Pharmaceuticals for Human (ICH), Harmonised Tripartite
Guideline, Evaluation of Stability Data Q1E, step 4, February 2003
(CPMP/ICH/420/02).\cr
\url{https://database.ich.org/sites/default/files/Q1E\%20Guideline.pdf}
}
\seealso{
\code{\link{expirest_osle}}, \code{\link{find_poi}},
\code{\link[stats]{uniroot}}, \code{\link[stats]{lm}},
\code{\link[stats]{AIC}}.
}
