% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_hlines}
\alias{get_hlines}
\title{Prepare horizontal lines}
\usage{
get_hlines(sl, ivl_side)
}
\arguments{
\item{sl}{A numeric value or a numeric vector of length \code{2} that
specifies the specification limit or limits.}

\item{ivl_side}{A character string that specifies if the \dQuote{upper} or
the \dQuote{lower} limit is the relevant limit, i.e. either \code{"upper"}
or \code{"lower"}, respectively.}
}
\value{
A data frame with the columns Response, Item, Colour and Type
is returned.
}
\description{
The function \code{get_hlines()} prepares a data frame for putting
horizontal lines on a plot prepared by the \code{ggplot()} function from
the \sQuote{\code{ggplot2}} package
}
\details{
The function \code{get_hlines()} expects various pieces
of information characterising an \sQuote{\code{expirest_osle}} or an
\sQuote{\code{expirest_wisle}} model. Based on the information provided,
the function prepares a data frame that can be handed over to the
\code{geom_hline()} function from the \sQuote{\code{ggplot2}} package.
}
\seealso{
\code{\link{plot_expirest_osle}}, \code{\link{plot_expirest_wisle}},
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_text}}.
}
\keyword{internal}
