% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_segments}
\alias{get_segments}
\title{Prepare segments explaining graphical elements}
\usage{
get_segments(
  sl,
  ivl_side,
  t_exp,
  rl,
  rl_index,
  poi_woca,
  wc_icpt,
  x_range,
  sl_model_name,
  wcsl_model_name
)
}
\arguments{
\item{sl}{A numeric value or a numeric vector of length \code{2} specifying
the specification limit or limits.}

\item{ivl_side}{A character string specifying if the \dQuote{upper} or the
\dQuote{lower} limit is the relevant limit, i.e. either \code{"upper"} or
\code{"lower"}, respectively.}

\item{t_exp}{A data frame of the intercepts, the differences between release
and shelf life limits, the worst case scenario limits (WCSLs), the expiry
and release specification limits, the shelf lives and POI values. The
default is \code{NULL}.}

\item{rl}{A numeric value or a numeric vector specifying the release
specification limit(s) for which the corresponding expiry should be
estimated.}

\item{rl_index}{A positive integer specifying which of the release limit
values that have been handed over to \code{\link{expirest_wisle}()} should
be displayed. The default is \code{NULL}.}

\item{poi_woca}{Point of intersection (POI) with the upper or lower
confidence or prediction interval of the linear regression model
representing the worst case scenario (woca) model. The default is
\code{NULL}.}

\item{wc_icpt}{A data frame of the worst case intercepts of each of the
four fitted models. The default is \code{NULL}.}

\item{x_range}{A numeric vector of the form \code{c(min, max)} specifying
the range of the time variable to be plotted.}

\item{sl_model_name}{A character string specifying the name of the
model with the shelf life limit.}

\item{wcsl_model_name}{A character string specifying the name of the
model with the worst case scenario limit. The default is \code{NULL}.}
}
\value{
A data frame with the columns Time.1, Time.2, Response.1,
Response.2, Item, Colour, Type, Size. Time.1 and Time.2 represent the
maximal allowed difference over time from intercept and Response.1 and
Response.2 represent the release limit.
}
\description{
The function \code{get_segments()} prepares a data frame for putting
segments on a plot prepared by the \code{ggplot()} function from
the \sQuote{\code{ggplot2}} package
}
\details{
The function \code{get_segments()} expects various pieces
of information characterising an \sQuote{\code{expirest_osle}} or an
\sQuote{\code{expirest_wisle}} model. Based on the information provided,
the function prepares a data frame that can be handed over to the
\code{geom_vline()} function from the \sQuote{\code{ggplot2}} package.
}
\seealso{
\code{\link{plot_expirest_osle}}, \code{\link{plot_expirest_wisle}},
  \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_text}}.
}
\keyword{internal}
