% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_text_annotation}
\alias{get_text_annotation}
\title{Prepare text annotation}
\usage{
get_text_annotation(
  rvu,
  x_range,
  y_range,
  sl,
  sl_sf,
  poi_model,
  ivl_side,
  poi_woca = NULL,
  t_exp = NULL,
  wc_icpt = NULL,
  rl_sf = NULL,
  rl_index = NULL,
  wcsl_model_name = NULL,
  plot_option = "full"
)
}
\arguments{
\item{rvu}{A character string specifying the unit associated with the
response variable.}

\item{x_range}{A numeric vector of the form \code{c(min, max)} specifying
the range of the time variable to be plotted.}

\item{y_range}{A numeric vector of the form \code{c(min, max)} specifying
the range of the response variable to be plotted.}

\item{sl}{A numeric value or a numeric vector of length \code{2} specifying
the specification limit or limits.}

\item{sl_sf}{A positive integer or a vector of positive integers specifying
the number of \dQuote{significant figures} (sf) of \code{sl}.}

\item{poi_model}{Point of intersection (POI) with the upper or lower
confidence or prediction interval of the most appropriate model according
to the ICH Q1E guideline.}

\item{ivl_side}{A character string specifying if the \dQuote{upper} or the
\dQuote{lower} limit is the relevant limit, i.e. either \code{"upper"} or
\code{"lower"}, respectively.}

\item{poi_woca}{Point of intersection (POI) with the upper or lower
confidence or prediction interval of the linear regression model
representing the worst case scenario (woca) model. The default is
\code{NULL}.}

\item{t_exp}{A data frame of the intercepts, the differences between release
and shelf life limits, the worst case scenario limits (WCSLs), the expiry
and release specification limits, the shelf lives and POI values. The
default is \code{NULL}.}

\item{wc_icpt}{A data frame of the worst case intercepts of each of the
four fitted models. The default is \code{NULL}.}

\item{rl_sf}{A positive integer or a vector of positive integers specifying
the number of \dQuote{significant figures} (sf) of \code{rl}. The default
is \code{NULL}.}

\item{rl_index}{A positive integer specifying which of the release limit
values that have been handed over to \code{\link{expirest_wisle}()} should
be displayed. The default is \code{NULL}.}

\item{wcsl_model_name}{A character string specifying the name of the
model with the worst case scenario limit. The default is \code{NULL}.}

\item{plot_option}{A character string of either \code{"full"},
\code{"lean1"}, \code{"lean2"}, \code{"basic1"} and \code{"basic2"},
specifying if additional information should be shown in the plot (option
\code{"full"}) or only basic information (options \code{"lean"} and
\code{"basic"}). Full means the data points, the fitted regression line
with the confidence or prediction interval, the specification limit(s)
and the estimated shelf life. The default is \code{"full"}.}
}
\value{
A data frame with the columns Time, Response, Label and Colour
is returned.
}
\description{
The function \code{get_text_annotation()} prepares a data frame for putting
text on a plot prepared by the \code{ggplot()} function from
the \sQuote{\code{ggplot2}} package.
}
\details{
The function \code{get_text_annotation()} expects various pieces
of information characterising an \sQuote{\code{expirest_osle}} or an
\sQuote{\code{expirest_wisle}} model. Based on the information provided,
the function prepares a data frame that can be handed over to the
\code{geom_text()} function from the \sQuote{\code{ggplot2}} package.
}
\seealso{
\code{\link{plot_expirest_osle}}, \code{\link{plot_expirest_wisle}},
  \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_text}}.
}
\keyword{internal}
