% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_wcs_limit}
\alias{get_wcs_limit}
\title{Determination of the \dQuote{worst case scenario} (wcs) limit}
\usage{
get_wcs_limit(
  rl,
  sl,
  intercept,
  xform = c("no", "no"),
  shift = c(0, 0),
  ivl_side = "lower"
)
}
\arguments{
\item{rl}{A numeric value specifying the release specification limit(s),
on the same scale as \code{sl} and \code{intercept}.}

\item{sl}{A numeric value specifying the specification limit, on the same
scale as \code{rl} and \code{intercept}.}

\item{intercept}{A numeric value representing the intercept of a linear
regression model fitted to sample data, on the same scale as \code{rl}
and \code{sl}.}

\item{xform}{A vector of two character strings specifying the transformation
of the response and the time variable. The default is \dQuote{no}
transformation, i.e. \code{c("no", "no")}, where the first element
specifies the transformation of the \eqn{x} variable and the second
element the transformation of the \eqn{y} variable. Valid alternatives
for \eqn{x} and/or \eqn{y} variable transformation are \code{"log"}
(natural logarithm), \code{"sqrt"} (square root) and \code{"sq"} (square).}

\item{shift}{A vector of two values which will be added to the variables
\eqn{x} and/or \eqn{y} before they are transformed as specified by the
\code{xform} parameter, where the first element will be added to the
\eqn{x} variable and the second element to the \eqn{y} variable. The
purpose is to prevent an undefined state which could arise when variables
with values of \eqn{\leq 0} are log or square root transformed. The
default is \code{c(0, 0)}. For transformation of the \eqn{x} variable
\eqn{1} should be chosen as shift parameter in case of \code{log} and
\eqn{0} in case of \code{sqrt} and \code{sq} transformation because
then the intercept will be at \eqn{x = 0}. For transformation of the
\eqn{y} variable any value can be chosen as shift parameter.}

\item{ivl_side}{A character string specifying if the \dQuote{upper} or the
\dQuote{lower} limit is the relevant limit, i.e. either \code{"upper"} or
\code{"lower"}, respectively. The default is \code{"lower"}.}
}
\value{
A list with the following elements is returned:
\item{xform}{A vector of two character strings specifying the transformation
  of the response and the time variable.}
\item{shift}{A vector of two values which has been added to the values of
  the transformed \eqn{x} and/or \eqn{y} variables (specified via the
  \code{xform} parameter).}
\item{wcsl}{A numeric value or a numeric vector of the worst case scenario
  limit(s) (wcsl) on the original scale.}
\item{wcsl.trfmd}{An optional item containing the numeric value or a numeric
  vector of the worst case scenario limit(s) (wcsl) on the transformed
  scale.}
}
\description{
The function \code{get_wcs_limit()} calculates \dQuote{worst case scenario}
(wcs) limit following the ARGPM Guidance \dQuote{Stability testing for
prescription medicines}.
}
\details{
The function \code{get_wcs_limit()} determines the \dQuote{worst
case scenario} (wcs) limit as is proposed by the Australian Regulatory
Guidelines for Prescription Medicines (ARGPM) guidance \dQuote{Stability
testing for prescription medicines}. According to this guideline, the shelf
life or expiry limit is estimated as the point where the upper or lower
limit of the 95\% confidence interval of the linear model fitted to the
data intersects the wcs limit. The wcs limit is obtained by
adding/subtracting the absolute difference of specification limit and the
release limit to/from the common intercept of the test batches or the
intercept of the worst performing batch.

If data have been linearised by transformation, all elements, i.e. \code{rl},
\code{sl} and \code{intercept} must be on the same, i.e. transformed, scale.
The results are returned on the transformed scale and also on the original
scale.
}
\references{
Therapeutic Goods Administration (TGA) of the Department of Health of the
Australian Government, Australian Regulatory Guidelines for Prescription
Medicines (ARGPM), Stability testing for prescription medicines,
Version 1.1, March 2017 \cr
\url{https://www.tga.gov.au/stability-testing-prescription-medicines}
}
\seealso{
\code{\link{expirest_wisle}}, \code{\link[stats]{lm}}.
}
\keyword{internal}
