% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.expirest_osle}
\alias{print.expirest_osle}
\title{Print a summary of the shelf life estimation (osle)}
\usage{
\method{print}{expirest_osle}(x, ...)
}
\arguments{
\item{x}{An \sQuote{\code{expirest_osle}} object, i.e. a list returned by
the \code{expirest_osle()} function.}

\item{...}{Further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{formatC}()} function.}
}
\value{
The \sQuote{\code{expirest_osle}} object passed to the
\code{object} parameter is returned invisibly.
}
\description{
This is a method for the function \code{print()} for objects of class
\sQuote{\code{expirest_osle}}.
}
\details{
The function \code{print.expirest_osle()} prints the most relevant
information in an \sQuote{\code{expirest_osle}} object.
}
\seealso{
\code{\link{expirest_osle}}, \code{\link{expirest_wisle}},
\code{\link[base]{formatC}}, \code{\link[utils]{methods}}.
}
