% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{summary.expirest_wisle}
\alias{summary.expirest_wisle}
\title{Summary of the what-if shelf life estimation (wisle)}
\usage{
\method{summary}{expirest_wisle}(object, ...)
}
\arguments{
\item{object}{an \code{expirest_wisle} object, i.e. a list returned by the
\sQuote{expirest_wisle()} function.}

\item{...}{further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{formatC}} function.}
}
\value{
The \code{expirest_wisle} object passed to the \code{model} parameter
is returned invisibly.
}
\description{
This is a method for the function \code{summary()} for objects of class
\code{expirest_wisle}.
}
\details{
The function \code{summary.expirest_wisle()} prints the most
relevant information in an \code{expirest_wisle} object.
}
\seealso{
\code{\link{expirest_wisle}}, \code{\link{expirest_osle}},
\code{\link[base]{formatC}}, \code{\link[utils]{methods}}.
}
